fixed unit test
